refactor: 랜딩페이지 컴포넌트 분리
