+++ frontend/src/components/landing/LoginModal/LoginModal.tsx
@@ -13,7 +13,7 @@ import GoogleIcon from '@Assets/icons/GoogleIcon';
 
 const REDIRECT_URI_PARAMETER = '/auth?provider=google';
 
-const Login = () => {
+const LoginModal = () => {
   const baseUri = `${window.location.protocol}//${window.location.host}`;
 
   const googleOAuthUrl = `https://accounts.google.com/o/oauth2/v2/auth?scope=https://www.googleapis.com/auth/userinfo.profile https://www.googleapis.com/auth/userinfo.email&client_id=${process.env.REACT_APP_GOOGLE_CLIENT_ID}&response_type=code&redirect_uri=${baseUri}${REDIRECT_URI_PARAMETER}`;
@@ -51,7 +51,7 @@ const Login = () => {
   );
 };
 
-export default Login;
+export default LoginModal;
 
 const Layout = styled.div`
   display: flex;
